Strengthen pattern pipeline verification and remove quadratic Core environment copies; add Haskell frontend, Core/CorePrep wire, Xpp/Xmm, LLVM, AARC, analyzer, formatter and linter regressions; correct uint type identities and unique Xmm diagnostics; make Criterion output Unicode-safe; document the measured linear verifier result and retain Catch2 after capability review

+++ Compiler/Haskell/Core/src/Visual/XSharp/Desugarer.hs
@@ -154,7 +154,7 @@ lowerPattern subject subjectType patternValue = case patternValue of
     TypePattern _ _ targetType ->
         CorePrimitive
             CoreTypeIs
-            [subject, CoreLiteral (CoreInteger (toInteger (typeIdentity targetType))) (namedType "ulong")]
+            [subject, CoreLiteral (CoreInteger (toInteger (typeIdentity targetType))) (namedType "uint")]
             boolType
     RelationalPattern _ operator literal literalType ->
         CorePrimitive
